+++ src/test/java/StringAddCalculatorTest.java
@@ -1,4 +1,4 @@
-import StringCalc.StringAddCalculator;
+import stringcalculation.StringAddCalculator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -15,31 +15,31 @@ public class StringAddCalculatorTest {
     }
 
     @Test
-    public void splitAndSum_숫자하나() throws Exception {
+    public void splitAndSum_숫자하나() {
         int result = StringAddCalculator.splitAndSum("1");
         assertThat(result).isEqualTo(1);
     }
 
     @Test
-    public void splitAndSum_쉼표구분자() throws Exception {
+    public void splitAndSum_쉼표구분자() {
         int result = StringAddCalculator.splitAndSum("1,2");
         assertThat(result).isEqualTo(3);
     }
 
     @Test
-    public void splitAndSum_쉼표_또는_콜론_구분자() throws Exception {
+    public void splitAndSum_쉼표_또는_콜론_구분자() {
         int result = StringAddCalculator.splitAndSum("1,2:3");
         assertThat(result).isEqualTo(6);
     }
 
     @Test
-    public void splitAndSum_custom_구분자() throws Exception {
+    public void splitAndSum_custom_구분자() {
         int result = StringAddCalculator.splitAndSum("//;\n1;2;3");
         assertThat(result).isEqualTo(6);
     }
 
     @Test
-    public void splitAndSum_negative() throws Exception {
+    public void splitAndSum_negative() {
         assertThatThrownBy(() -> StringAddCalculator.splitAndSum("-1,2,3"))
                 .isInstanceOf(RuntimeException.class);
     }
